Use new sender in ingest
