+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/runner/impl/commit/AddFilesToStateStore.java
@@ -15,28 +15,23 @@
  */
 package sleeper.ingest.runner.impl.commit;
 
-import com.amazonaws.services.sqs.AmazonSQS;
-import com.amazonaws.services.sqs.model.SendMessageRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.statestore.AllReferencesToAFile;
 import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.core.statestore.commit.StateStoreCommitRequest;
-import sleeper.core.statestore.commit.StateStoreCommitRequestUploader;
+import sleeper.core.statestore.commit.StateStoreCommitRequestSender;
 import sleeper.core.statestore.transactionlog.transactions.AddFilesTransaction;
 import sleeper.core.tracker.ingest.job.IngestJobTracker;
 import sleeper.core.tracker.ingest.job.update.IngestJobAddedFilesEvent;
 
 import java.util.List;
-import java.util.UUID;
 import java.util.function.Consumer;
 
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.STATESTORE_COMMITTER_QUEUE_URL;
 import static sleeper.core.properties.table.TableProperty.TABLE_ID;
 
 @FunctionalInterface
@@ -59,22 +54,15 @@ public interface AddFilesToStateStore {
     }
 
     static AddFilesToStateStore bySqs(
-            AmazonSQS sqsClient, InstanceProperties instanceProperties, TableProperties tableProperties,
-            StateStoreCommitRequestUploader uploader,
+            TableProperties tableProperties, StateStoreCommitRequestSender commitSender,
             Consumer<AddFilesTransaction.Builder> transactionConfig) {
         return references -> {
-            AddFilesTransaction.Builder requestBuilder = AddFilesTransaction.builder()
-                    .files(AllReferencesToAFile.newFilesWithReferences(references));
+            List<AllReferencesToAFile> files = AllReferencesToAFile.newFilesWithReferences(references);
+            AddFilesTransaction.Builder requestBuilder = AddFilesTransaction.builder().files(files);
             transactionConfig.accept(requestBuilder);
             AddFilesTransaction transaction = requestBuilder.build();
-            StateStoreCommitRequest request = StateStoreCommitRequest.create(tableProperties.get(TABLE_ID), transaction);
-            LOGGER.debug("Sending asynchronous request to state store committer: {}", request);
-            sqsClient.sendMessage(new SendMessageRequest()
-                    .withQueueUrl(instanceProperties.get(STATESTORE_COMMITTER_QUEUE_URL))
-                    .withMessageBody(uploader.serialiseAndUploadIfTooBig(request))
-                    .withMessageGroupId(request.getTableId())
-                    .withMessageDeduplicationId(UUID.randomUUID().toString()));
-            LOGGER.info("Submitted asynchronous request to add files via state store committer queue");
+            commitSender.send(StateStoreCommitRequest.create(tableProperties.get(TABLE_ID), transaction));
+            LOGGER.info("Submitted asynchronous request to add {} files with {} references", files.size(), references.size());
         };
     }
 }
